requirements.txt
discord.py
discord.py
ffmpeg
youtube_dl
pandas
numpy
matplotlib